Switched back to JAudioTagger, thus switched back to Java 1.8 - with necessary fixes and some refactorings. Fixed playlist jumping on play with search text entered.

+++ src/main/java/x/mvmn/sonivm/audio/AudioFileInfo.java
@@ -8,6 +8,6 @@ import lombok.Data;
 public class AudioFileInfo {
 
 	private final String filePath;
-	private final Long durationSeconds;
+	private final Integer durationSeconds;
 	private final boolean seekable;
 }
